requirements.txt
django==4.2
djangorestframework==3.15
celery==5.3
redis==5.0
motor==3.3
httpx==0.27
python-dotenv==1.0
django-environ==0.11
pymongo==4.6
structlog==24.1